dialogflow_api_call/requirements.txt
google-cloud-dialogflow-cx
google-cloud-texttospeech>=2.29.0
python-dotenv
pyaudio
sounddevice
soundfile
